Fix Devin review #581 (BUG-1, INFO-1): restore TestUpdateAbuseUserReport + reorg
🟡 BUG-1: 範囲削除で TestUpdateAbuseUserReport も巻き込んでしまった
(handler は handler_stubs.go に残ってるのでテストだけ消えると 80% 例外
閾値を割る可能性)。moderation_test.go に復元。本 endpoint は abuse
report の resolved フラグを立てる moderation 操作なので moderation 配置
が妥当。

📝 INFO-1: TestDeleteAllFilesOfUser_DeletesOnlyTargetUserFiles を
accounts_test.go から moderation_test.go に移動。handler が
moderation.go にあるので discoverability の整合性改善。

残り 2 件 (英日混在 docstring / mock case 確認) は info でポリシーに従い
修正なし。
